requirements.txt
Django==3.2.8
pandas==1.3.4
scikit-learn==0.24.2
django-formtools==2.4
django-bootstrap4
